Stopmotion Changes
Didnt finish, but:
Cropped image to size so black line wasnt in it of Mai in window.
Fixed sizing of rotating images on stopmotion page.
Removed titles from the videos, because they were on the vimeo links.
In Progress:
Attempting to make all of the images have 20px margin on each side, and have them all be the same size. didnt finish yet will continue later.